% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprodV.R
\name{crossprodV}
\alias{crossprodV}
\title{Matrix Inner Product}
\usage{
crossprodV(X, V, inverse = FALSE)
}
\arguments{
\item{X}{Either a matrix of size \code{p x q}, or an array of size \code{p x q x n}.}

\item{V}{Either a matrix of size \code{p x q}, or an array of size \code{p x q x n}.}

\item{inverse}{Whether or not the inner product should be calculated with \code{V} (\code{inverse = FALSE}) or \code{V^{-1}}.}
}
\value{
An array of size \code{q x q x n}.
}
\description{
Vectorized matrix inner products \code{Y = X' V X} or \code{Y = X' V^{-1} X}.
}

