% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\docType{data}
\name{Hospitals}
\alias{Hospitals}
\title{Hospital profiling data.}
\format{A data frame with 27 rows (one for each hospital) and 4 variables:
\describe{
  \item{\code{NSrg}}{Non-surgery related problem rate (\%).}
  \item{\code{Srg}}{Surgery related problem rate (\%).}
  \item{\code{Severity}}{Average health index for surveyed patients.}
  \item{\code{Size}}{Number of patients surveyed.}
}}
\usage{
Hospitals
}
\description{
Information on patient-reported problem rates for 27 teaching hosptials and private academic health centers in the United States.
}
\references{
Everson & Morris, 2000; original data in Cleary et al, 1991.
}
\keyword{datasets}
