% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixT.R
\name{MatrixT}
\alias{MatrixT}
\alias{dMT}
\title{The Matrix-t distribution.}
\usage{
dMT(X, Mu, RowV, ColV, nu, log = FALSE)
}
\arguments{
\item{X}{argument to the density function.  Either a \code{p x q} matrix or a \code{p x q x n} array.}

\item{Mu}{mean matrix or matrices.  Either a \code{p x q} matrix or a \code{p x q x n} array.}

\item{RowV}{between-row covariance matrix or matrices.  Either a \code{p x p} matrix or a \code{p x p x n} array.}

\item{ColV}{between-column covariance matrix or matrices.  Either a \code{q x q} matrix or a \code{q x q x n} array.}

\item{nu}{degrees-of-freedom parameter.  A scalar or vector.}

\item{log}{logical. Whether or not to compute the log-density.}
}
\value{
A vector for densities, or a \code{p x q x n} array for random sampling.
}
\description{
Density and sampling for the Matrix-t distribution.
}
