% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mniw-package.R
\docType{package}
\name{mniw}
\alias{mniw}
\alias{mniw-package}
\title{Density and Simulation for the Matrix Normal Inverse Wishart (MNIW) Distribution.}
\description{
Density and Simulation for the Matrix Normal Inverse Wishart (MNIW) Distribution.
}
\details{
The Matrix-Normal Inverse-Wishart (MNIW) distribution on \code{(X,V)} is given by
\deqn{X | V ~ MN(Lambda, Sigma, V),}
\deqn{V ~ iWish(Psi, nu).}
The default parameters are \code{Lambda = matrix(0,p,q)}, \code{Sigma = diag(p)}, \code{Psi = diag(q)}.
}
