% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffects.R
\name{mNormRE.post}
\alias{mNormRE.post}
\title{Bayesian Inference for Multivariate Normal Random-Effects Model}
\usage{
mNormRE.post(nsamples, Y, V, X, prior = NULL, init, burn, updateHyp = TRUE,
  storeHyp = TRUE, updateRE = TRUE, storeRE = FALSE)
}
\arguments{
\item{nsamples}{Number of posterior samples to draw.}

\item{Y}{\code{N x q} matrix of responses.}

\item{V}{Either a \code{q x q} variance matrix or an \code{q x q x N} array of such matrices.}

\item{X}{\code{N x p} matrix of covariates.}

\item{prior}{Parameters of the prior MNIW distribution on the hyperparameters.  See Details.}

\item{init}{Optional list with elements \code{Beta}, \code{Sigma}, and \code{Mu} providing the initial values for these.  Default values are \code{Beta = matrix(0, p, q)}, \code{Sigma = diag(q)}, and \code{Mu = Y}.}

\item{burn}{Integer number of burn-in samples, or fraction of \code{nsamples} to prepend as burn-in.}

\item{updateHyp, storeHyp}{Whether or not to update/store the hyperparameter draws.}

\item{updateRE, storeRE}{Whether or not to update/store the random-effects draws.}
}
\value{
A list with potential elements:
\describe{
  \item{\code{Beta}}{An \code{p x q x nsamples} array of regression coefficient iterations (if \code{storeHyp == TRUE})}
  \item{\code{Sigma}}{An \code{q x q x nsamples} array of regression variance matrices (if \code{storeHyp == TRUE})}
  \item{\code{Mu}}{An \code{n x q x nsamples} array of random effects (if \code{storeRE == TRUE})}
}
}
\description{
Gibbs sampler for posterior distribution of parameters and hyperparameters of the Multivariate Normal Random Effects (mNormRE) model.
}
\details{
The mNormRE model is given by
\deqn{
y_i \mid \mu_i \sim_iid N(\mu_i, V_i)
}
\deqn{
\mu_i \mid \beta, \Sigma ~sim_ind N(x_i' \beta, \Sigma)
}
\deqn{
\beta, \Sigma \sim MNIW(\Lambda, \Omega^{-1}, \Psi, \nu),
}
where \eqn{y_i} and \eqn{\mu_i} are response and random-effects vectors of length \eqn{q}, \eqn{x_i} are covariate vectors of length \eqn{p}, and \eqn{(\beta, \Sigma)} are hyperparameter matrices of size \eqn{p \times q} and \eqn{\q \times q}.

The MNIW prior distribution is given by a list with elements \code{Lambda}, \code{Omega}, \code{Psi}, and \code{nu}.  If any of these is \code{NULL} or missing, the default value is 0.  Note that \code{Omega == 0} gives a Lebesgue prior to \eqn{\beta}.
}
