% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiNormal.R
\name{MultiNormal}
\alias{MultiNormal}
\alias{dmNorm}
\alias{rmNorm}
\alias{rmNorm}
\title{The Multivariate Normal Distribution.}
\usage{
dmNorm(x, mu, V, log = FALSE)

rmNorm(n, mu, V, debug = FALSE)
}
\arguments{
\item{x}{Argument to the density function.  A vector of length \code{q} or an \code{n x q} matrix.}

\item{mu}{Mean vector(s).  Either a vector of length \code{q} or an \code{n x q} matrix.}

\item{V}{Covariance matrix or matrices.  Either a \code{q x q} matrix or a \code{q x q x n} array.}

\item{log}{Whether or not to compute the log-density.}

\item{n}{Number of random vectors to generate.}
}
\value{
A vector for densities, or a \code{n x q} array for random sampling.
}
\description{
Density and Random sampling for the Multivariate Normal Distribution.
}
\details{
\code{dmNorm} and \code{rmNorm} both accept single or multiple values for each argument.
}
