% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNIW.R
\name{rMNIW}
\alias{rMNIW}
\title{Generate samples from the Matrix-Normal Inverse-Wishart distribution.}
\usage{
rMNIW(n, Lambda, Sigma, Psi, nu, prec = FALSE)
}
\arguments{
\item{n}{number of samples.}

\item{Lambda}{\code{p x q} mean matrix of \code{X} (see details).}

\item{Sigma}{\code{p x p} conditional row variance matrix of \code{X}.}

\item{Psi}{\code{q x q} scale matrix of \code{V}.}

\item{nu}{scalar degrees of freedom of \code{V}.}

\item{prec}{logical. Whether or not to return \code{V} or \code{C = V^{-1}}.}
}
\value{
A list with the folllwing elements:
\describe{
\item{X}{Random samples for X ~ MN(Lambda, Sigma, V).}
\item{V}{Random samples for V ~ iWish(Psi, nu).}
}
}
\description{
Generate samples from the Matrix-Normal Inverse-Wishart distribution.
}
\details{
The Matrix-Normal Inverse-Wishart (MNIW) distribution on \code{p x q} matrix \code{X} and \code{q x q} matrix \code{V} is:
\deqn{X | V ~ MN(Lambda, Sigma, V),}
\deqn{V ~ iWish(Psi, nu).}
}
\examples{
## Sample from MNIW distribution
n = 100
p = 2
q = 2
Lambda = matrix(c(1,-0.5,-1,1),p,q)
Sigma = matrix(c(1,0.5,0.5,1),p,p)
Psi = matrix(c(1,0.1,0.1,1),p,q)
nu = q + 1
rMNIW(n, Lambda, Sigma, Psi, nu)

}
