% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffects.R
\name{rmNormRE}
\alias{rmNormRE}
\title{Conditional sampling for Multivariate Normal Random-Effects model.}
\usage{
rmNormRE(n, y, V, lambda, A)
}
\arguments{
\item{n}{the number of random samples to generate.}

\item{y}{vector of length \code{q} or \code{n x q} matrix of observations.  In the latter case each row is a different vector (see Details).}

\item{V}{matrix of size \code{q x q} or \code{q x q x n} array of observation variances.}

\item{lambda}{vector of length \code{q} or \code{n x q} matrix of prior means.  In the latter case each row is a different mean.  Defaults to zeros.}

\item{A}{matrix of size \code{q x q} or \code{q x q x n} array of prior variances.  Defaults to identity matrix.}
}
\description{
Sample from the conditional parameter distribution given the data and hyperparameters of the Multivariate Normal Random Effects (mNormRE) model.
}
\details{
The normal-normal random effects model is:
\deqn{y | \mu \sim N(\mu, V),}
\deqn{\mu \sim N(\lambda, A).}
The posterior distribution for the random effects \eqn{\mu} is
\deqn{\mu | y \sim N(B(\lambda - y) + y,(I - B)V),}
where \eqn{B = V(V+A)^{-1}}.
}
\examples{
# Conditional sampling with default prior
n = 100
q = 10
y = rnorm(q)
V = rwish(1, diag(q), q+1)
lambda = rep(0,q)
A = diag(q)
rmNormRE(n, y, V, lambda, A)

}
