% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wishart.R
\name{Wishart}
\alias{Wishart}
\alias{dwish}
\alias{rwish}
\alias{diwish}
\alias{riwish}
\alias{dwishart}
\alias{rwishart}
\alias{rwish}
\alias{diwish}
\alias{riwish}
\alias{dwishart}
\alias{rwishart}
\title{Wishart and Inverse-Wishart Distributions.}
\usage{
dwish(X, Psi, nu, log = FALSE)

rwish(n, Psi, nu)

diwish(X, Psi, nu, log = FALSE)

riwish(n, Psi, nu)

dwishart(X, Psi, nu, inverse = FALSE, log = FALSE, debug = FALSE)

rwishart(n, Psi, nu, inverse = FALSE)
}
\arguments{
\item{X}{Argument to the density function.  Either a \code{q x q} matrix or a \code{q x q x n} array.}

\item{Psi}{Scale parameter.  Either a \code{q x q} matrix or a \code{q x q x n} array.}

\item{nu}{Degrees-of-freedom parameter.  A scalar or vector.}

\item{log}{Whether or not to compute the log-density.}

\item{n}{Number of random matrices to generate.}

\item{inverse}{Whether or not to use the Inverse-Wishart distribution.}
}
\value{
A vector for densities, or a \code{q x q x n} array for random sampling.
}
\description{
Densities and Random Sampling for the Wishart and Inverse-Wishart distributions.
}
\details{
\code{dwish} and \code{diwish} are convenience wrappers for \code{dwishart}, and similarly \code{rwish} and \code{riwish} are wrappers for \code{rwishart}.

The distribution of a \code{q x q} Wishart random matrix is
\deqn{\code{f(X) = const |X|^{(\nu-q-1)/2} * \exp{-0.5 * Tr(\Psi^{-1} X)}},}
where \code{Psi} is a symmetric positive-definite matrix and \code{nu > q-1}.

The distribution of an Inverse-Wishart random matrix is
\deqn{\code{f(X) \propto |X|^{-(\nu+q+1)/2} * \exp{-0.5 * Tr(\Psi X^{-1})}}.}
If \code{X ~ Wish(Psi, nu)}, then \code{X^{-1} ~ iWish(Psi^{-1}, nu)}.

Quadratic forms involving a nonrandom vector \code{a} and a Wishart random matrix \code{X} have a chi-square distribution.  That is, for fixed \code{a} and \code{X ~ Wish(Psi, nu)},
\deqn{\code{(a' X a) / (a' Psi a) ~ chisq(df = nu)}.}
}
