% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprodV.R
\name{crossprodV}
\alias{crossprodV}
\title{Matrix cross-product.}
\usage{
crossprodV(X, Y = NULL, V, inverse = FALSE)
}
\arguments{
\item{X}{either: a matrix of size \code{p x q}, or an array of size \code{p x q x n}.}

\item{Y}{(optional), either: a matrix of size \code{p x r}, or an array of size \code{p x r x n}.  If missing uses \code{Y = X}.}

\item{V}{either: a matrix of size \code{p x p}, or an array of size \code{p x p x n}.}

\item{inverse}{logical. Indicates whether or not the inner product should be calculated with \code{V} or \code{V^{-1}}.}
}
\value{
An array of size \code{q x r x n}.
}
\description{
Vectorized matrix cross-products \code{t(X) V Y} or \code{t(X) V^{-1} Y}.
}
\examples{
## Basic vectorized cross-product
p = 4
q = 2
X = matrix(rnorm(p*q),p,q)
V = diag(p)
crossprodV(X=X,V=V)

}
