% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixNormal.R
\name{MatrixNormal}
\alias{MatrixNormal}
\alias{dMNorm}
\alias{rMNorm}
\alias{rMNorm}
\title{Matrix-Normal Distribution.}
\usage{
dMNorm(X, Mu, RowV, ColV, log = FALSE, debug = FALSE)

rMNorm(n, Mu, RowV, ColV, debug = FALSE)
}
\arguments{
\item{X}{Argument to the density function.  Either a \code{p x q} matrix or a \code{p x q x n} array.}

\item{Mu}{Mean matrix or matrices.  Either a \code{p x q} matrix or a \code{p x q x n} array.}

\item{RowV}{Between-Row covariance matrix or matrices.  Either a \code{p x p} matrix or a \code{p x p x n} array.}

\item{ColV}{Between-Column covariance matrix or matrices.  Either a \code{q x q} matrix or a \code{q x q x n} array.}

\item{log}{Whether or not to compute the log-density.}

\item{n}{Number of random matrices to generate.}
}
\value{
A vector for densities, or a \code{p x q x n} array for random sampling.
}
\description{
Density and Random sampling for the Matrix-Normal Distribution.
}
\details{
\code{dMNorm} and \code{rMNorm} both accept single or multiple values for each argument.

The distribution of a \code{p x q} Matrix-Normal random matrix is
\deqn{f(X) = const * \exp{-0.5 Tr( ColV^{-1}(X-Mu)''RowV^{-1}(X-Mu) )}.}
Linear combinations of Matrix-Normals are normal.  That is, for fixed \code{a} and \code{b} and \code{X ~ MNorm(Mu, RowV, ColV)}, then
\deqn{a' X b \sim N( a' Mu b, (a' RowV a) * (b' ColV b) ).}
}
