% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiNormal.R
\name{MultiNormal}
\alias{MultiNormal}
\alias{dmNorm}
\alias{rmNorm}
\title{The Multivariate Normal distribution.}
\usage{
dmNorm(x, mu, V, log = FALSE)

rmNorm(n, mu, V)
}
\arguments{
\item{x}{argument to the density function.  A vector of length \code{q} or an \code{n x q} matrix.}

\item{mu}{mean vector(s).  Either a vector of length \code{q} or an \code{n x q} matrix.}

\item{V}{covariance matrix or matrices.  Either a \code{q x q} matrix or a \code{q x q x n} array.}

\item{log}{logical. Whether or not to compute the log-density.}

\item{n}{number of random vectors to generate.}
}
\value{
A vector for densities, or a \code{n x q} array for random sampling.
}
\description{
Density and random sampling for the Multivariate Normal distribution.
}
\details{
\code{dmNorm} and \code{rmNorm} both accept single or multiple values for each argument.
}
\examples{
## Multivariate Normal random sample and subsequent density calculation
n = 100
q = 2
mu = c(1,2)
V = toeplitz(exp(-seq(1:q)))

# Random sample from Multivariate Normal distribution
mnormData = rmNorm(n, mu, V)

# Calculate log density for each sampled vector
dmNorm(mnormData, mu, V, log=TRUE)
}
