% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNIW.R
\name{rMNIW}
\alias{rMNIW}
\title{Generate Samples from the Matrix-Normal Inverse-Wishart Distribution.}
\usage{
rMNIW(n, Lambda, Sigma, Psi, nu, prec = FALSE, debug = FALSE)
}
\arguments{
\item{n}{Number of samples.}

\item{Lambda}{\code{p x q} mean matrix of \code{X} (see details).}

\item{Sigma}{\code{p x p} conditional row variance matrix of \code{X}.}

\item{Psi}{\code{q x q} Precision matrix of \code{V}.}

\item{nu}{Scalar degrees of freedom of \code{V}.}

\item{prec}{Logical, whether or not to return \code{V} or \code{C = V^{-1}}.}
}
\details{
the Matrix-Normal Inverse-Wishart (MNIW) distribution on \code{p x q} matrix \code{X} and \code{q x q} matrix \code{V} is:
\deqn{X | V \sim MN(Lambda, Sigma, V), \qquad V \sim iWish(Psi, nu).}
}

