% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-effects.R
\name{rmNormRE}
\alias{rmNormRE}
\title{Random Effects Sampling for Multivariate Normal-Normal Model.}
\usage{
rmNormRE(n, y, V, lambda, A)
}
\arguments{
\item{n}{The number of random samples to generate.}

\item{y}{Vector of length \code{q} or \code{q x n} matrix of observations.  In the latter case each column is a different vector (see details).}

\item{V}{Matrix of size \code{q x q} or \code{q x q x n} array of observation variances.}

\item{lambda}{Vector of length \code{q} or \code{q x n} matrix of prior means.}

\item{A}{Matrix of size \code{q x q} or \code{q x q x n} array of prior variances.}
}
\description{
Sample from the multivariate normal conditional distribution of the random effects given the data.
}
\details{
The normal-normal random effects model is:
\deqn{y | mu \sim N(mu, V), \qquad mu \sim \N(lambda, A).}
The posterior distribution for the random effects \code{mu} is
\deqn{mu | y \sim N(,).}
}

