% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomEffects.R
\name{rmNormRE}
\alias{rmNormRE}
\title{Conditional Sampling for Multivariate Normal Random-Effects Model.}
\usage{
rmNormRE(n, y, V, lambda, A)
}
\arguments{
\item{n}{The number of random samples to generate.}

\item{y}{Vector of length \code{q} or \code{n x q} matrix of observations.  In the latter case each row is a different vector (see details).}

\item{V}{Matrix of size \code{q x q} or \code{q x q x n} array of observation variances.}

\item{lambda}{Vector of length \code{q} or \code{n x q} matrix of prior means.  In the latter case each row is a different mean.  Defaults to zeros.}

\item{A}{Matrix of size \code{q x q} or \code{q x q x n} array of prior variances.  Defaults to identity matrix.}
}
\description{
Sample from the conditional parameter distribution given the data and hyperparameters of the Multivariate Normal Random Effects (mNormRE) model.
}
\details{
The normal-normal random effects model is:
\deqn{y | mu \sim N(mu, V), \qquad mu \sim \N(lambda, A).}
The posterior distribution for the random effects \code{mu} is
\deqn{mu | y \sim N(,).}
}
